Move constraint2 to constraint
