+++ compiler-core/checking/src/core/constraint/compiler/prim_symbol.rs
@@ -4,8 +4,8 @@ use building_types::QueryResult;
 
 use crate::ExternalQueries;
 use crate::context::CheckContext;
-use crate::core::constraint2::WorkItem;
-use crate::core::constraint2::matching::{self, InstanceMatch, MatchInstance};
+use crate::core::constraint::WorkItem;
+use crate::core::constraint::matching::{self, InstanceMatch, MatchInstance};
 use crate::core::{Type, TypeId, normalise};
 use crate::state::CheckState;
